+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to the Gainium Backtester library will be documented in this
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.6] - 2026-01-12
+
+### Fixed 
+- Wrong combined profit hedge bot.
+- Wrong close time by combined settings. 
+
 ## [1.4.5] - 2026-01-06
 
 ### Fixed